refactor(Entity): Service Response 타입과 실 사용 Entity 타입을 명확히 구분하도록 수정
rename GenericList -> GenericResponse
rename CountryProvidersList -> ProvidersByCountryResponse
arrange files by corresponding groups
