+++ remy-movie/remy-movie/Data/Service/Movie/ProvidersByCountryResponse.swift
@@ -1,0 +1,20 @@
+//
+//  ProvidersByCountryResponse.swift
+//  remy-movie
+//
+//  Copyright (c) 2023 Jeremy All rights reserved.
+    
+
+import Foundation
+
+struct ProvidersByCountryResponse: Decodable {
+    let en: ProvidersInfo?
+    let kr: ProvidersInfo?
+    let jp: ProvidersInfo?
+    
+    enum CodingKeys: String, CodingKey {
+        case en = "EN"
+        case kr = "KR"
+        case jp = "JP"
+    }
+}
