+++ remy-movie/remy-movie/Domain/Entities/Keyword.swift
@@ -1,16 +1,11 @@
 //
-//  Keywords.swift
+//  Keyword.swift
 //  remy-movie
 //
 //  Copyright (c) 2023 Jeremy All rights reserved.
     
 
 import Foundation
-
-struct KeywordList: Decodable {
-    let id: Int?
-    let keywords: [Keyword]?
-}
 
 struct Keyword: Decodable {
     let id: Int?
